Consolidate featherreadmex, featherreadwritemex cc and h files into feather_functions.cc and feather_functions.h
