+++ matlab/src/cpp/arrow/matlab/feather/feather_functions.cc
@@ -17,12 +17,24 @@
 
 #include <string>
 
-#include "featherreadmex.h"
+#include "feather_writer.h"
 #include "feather_reader.h"
+#include "feather_functions.h"
 #include "util/handle_status.h"
 
-// MEX gateway function. This is the entry point for featherreadmex.cpp.
-void featherreadmex(int nlhs, mxArray* plhs[], int nrhs, const mxArray* prhs[]) {
+void featherwrite(int nlhs, mxArray* plhs[], int nrhs, const mxArray* prhs[]) {
+  const std::string filename{mxArrayToUTF8String(prhs[0])};
+
+  // Open a Feather file at the provided file path for writing.
+  std::shared_ptr<arrow::matlab::FeatherWriter> feather_writer{nullptr};
+  arrow::matlab::util::HandleStatus(
+      arrow::matlab::FeatherWriter::Open(filename, &feather_writer));
+
+  // Write the Feather file table variables and table metadata from MATLAB.
+  arrow::matlab::util::HandleStatus(feather_writer->WriteVariables(prhs[1], prhs[2]));
+}
+
+void featherread(int nlhs, mxArray* plhs[], int nrhs, const mxArray* prhs[]) {
   const std::string filename{mxArrayToUTF8String(prhs[0])};
 
   // Read the given Feather file into memory.
