+++ matlab/src/cpp/arrow/matlab/feather/feather_functions.h
@@ -18,3 +18,5 @@
 #include <mex.h>
 
 void featherwrite(int nlhs, mxArray* plhs[], int nrhs, const mxArray* prhs[]);
+
+void featherreadmex(int nlhs, mxArray* plhs[], int nrhs, const mxArray* prhs[]);
